aider: Add `--limit` option to process a specified number of subtitles.
# Aider chat conversation:

USER: Add `--limit` (optional). If None, the script will processs all subtitles. Otherwise, it will only process the # of subtitle specified by `--limit`.

ASSISTANT: To add the `--limit` option, we need to modify the `main.py` and `fromaudio.py` files.

In `main.py`, we need to add a new argument to the `fromaudio` parser.

In `fromaudio.py`, we need to modify the `fromaudio_main` function to pass the `args.limit` to the `create_audio_from_audio` function. Then, in the `create_audio_from_audio` function, we need to modify the loop that iterates over the `subtitle_data` to only process the number of subtitles specified by `limit`.

Here are the proposed changes:

<source>python
main.py
<<<<<<< HEAD
    parser_fromaudio.add_argument('--tr-strategy', default='deepl', help='Translation strategy to use. Options are "deepl" and "fake". Default is "deepl".')
=======
    parser_fromaudio.add_argument('--tr-strategy', default='deepl', help='Translation strategy to use. Options are "deepl" and "fake". Default is "deepl".')
    parser_fromaudio.add_argument('--limit', type=int, help='Number of subtitles to process. If not provided, all subtitles will be processed.')
>>>>>>> updated
</source>

<source>python
fromaudio.py
<<<<<<< HEAD
    create_audio_from_audio(
        args.input_audio,
        subtitle_data,
        args.output_file,
        args.transition_sound,
        args.repeat_count,
        args.tr_lang,
        args.verbose,
        translate_func=translate_func,
    )
=======
    create_audio_from_audio(
        args.input_audio,
        subtitle_data,
        args.output_file,
        args.transition_sound,
        args.repeat_count,
        args.tr_lang,
        args.verbose,
        translate_func=translate_func,
        limit=args.limit,
    )
>>>>>>> updated
</source>

<source>python
fromaudio.py
<<<<<<< HEAD
    for subtitle in tqdm(subtitle_data[280:], desc='Processing sentences'):
=======
    for subtitle in tqdm(subtitle_data[:limit], desc='Processing sentences'):
>>>>>>> updated
</source>